requirements.txt
torch
unstructured[local-inference]
detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2
layoutparser[layoutmodels,tesseract]
tensorboard>=2.12.2
langchain
discord
openai
tiktoken
chromadb
faiss-cpu
deepgram-sdk
peewee
playwright
lxml
requests
pytube
youtube-transcript-api
python-magic